Extract sysinfo logic to its own class.
